Canvas Slider & Smart Save
Added a canvas resizing slider at the bottom of the application window. Operates from -1 to 1, takes the value from the slider and adds 1, then uses that as a scaling factor.

Added smart save that detects if there have been unsaved changes since the last save. On closing a tab, will check for unsaved changes. If there are, it will not allow you to close the tab until you save. No behavior when closing the application!