Many Changes
Added a warning to clearing the canvas, fixed the warning for closing a tab.

Moved dashed widths to the width selector, replaced the dashed line button with the polygon button, currently connected to a non-functioning drawPolygon() method.

Added a StackPane to the tab hierarchy. New hierarchy is TabPane > Tab > ScrollPane > StackPane > Canvases.

To-Do:
Rest of sprint 4